scripts/requirements-ga4.txt
google-analytics-admin>=0.22.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1